Fix nightly warnings
Also move poly tests to its own test file since it tests the public API
