+++ secp256kfun/src/poly.rs
@@ -6,25 +6,15 @@
 //!
 //! [`Scalars`]: crate::Scalar
 //! [`Points`]: crate::Point
+use crate::{g, marker::*, s, Point, Scalar, G};
 #[cfg(feature = "alloc")]
 use alloc::vec::Vec;
-
-use crate::{
-    marker::{Public, Secrecy, ZeroChoice},
-    s, Scalar,
-};
+use core::iter;
+use rand_core::RngCore;
 
 /// Functions for dealing with scalar polynomials
 pub mod scalar {
     use super::*;
-    use rand_core::RngCore;
-
-    use crate::{
-        g,
-        marker::{Secrecy, Secret, Zero, ZeroChoice},
-        poly::powers,
-        s, Point, Scalar, G,
-    };
 
     /// Evaluate a scalar polynomial defined by coefficients, at some scalar index.
     ///
@@ -80,14 +70,6 @@ pub mod scalar {
 /// Functions for dealing with point polynomials
 pub mod point {
     use super::*;
-    use core::iter;
-
-    use crate::{
-        g,
-        marker::{NonNormal, PointType, Public, Secrecy, Zero, ZeroChoice},
-        poly::powers,
-        s, Point, Scalar,
-    };
 
     /// Evaluate a point polynomial defined by coefficients, at some index.
     ///
@@ -205,120 +187,4 @@ pub fn eval_basis_poly_at_0<'a>(
                 .expect("we filtered duplicate indicies");
             s!(acc * x_m / denominator).public()
         })
-}
-
-#[cfg(test)]
-mod test {
-    use crate::{
-        g,
-        marker::{Normal, Zero},
-        poly, Point, G,
-    };
-
-    use super::*;
-
-    #[test]
-    fn test_lagrange_lambda() {
-        let res = s!((1 * 4 * 5) / { s!((1 - 2) * (4 - 2) * (5 - 2)).non_zero().unwrap() });
-        assert_eq!(
-            res,
-            eval_basis_poly_at_0(s!(2), [s!(1), s!(4), s!(5)].iter())
-        );
-    }
-
-    #[test]
-    fn test_add_poly() {
-        let poly1 = vec![g!(1 * G), g!(2 * G), g!(3 * G)];
-        let poly2 = vec![g!(8 * G), g!(5 * G), g!(11 * G)];
-
-        let addition = poly::point::add(&poly1, &poly2);
-        assert_eq!(addition, vec![g!(9 * G), g!(7 * G), g!(14 * G)])
-    }
-
-    #[test]
-    fn test_add_poly_unequal_len() {
-        let poly1 = vec![g!(1 * G)];
-        let poly2 = vec![g!(8 * G), g!(5 * G)];
-        let addition = poly::point::add(&poly1, &poly2);
-        assert_eq!(addition, vec![g!(9 * G), g!(5 * G)]);
-
-        let poly1 = vec![g!(3 * G), g!(1 * G)];
-        let poly2 = vec![g!(5 * G)];
-        let addition = poly::point::add(&poly1, &poly2);
-        assert_eq!(addition, vec![g!(8 * G), g!(1 * G)]);
-    }
-
-    #[test]
-    fn test_recover_public_poly() {
-        let poly = vec![g!(1 * G), g!(2 * G), g!(3 * G)];
-        let indicies = vec![s!(1).public(), s!(3).public(), s!(2).public()];
-        let points = indicies
-            .clone()
-            .into_iter()
-            .map(|index| {
-                (
-                    index,
-                    poly::point::eval(&poly, index.public())
-                        .normalize()
-                        .non_zero()
-                        .unwrap(),
-                )
-            })
-            .collect::<Vec<_>>();
-
-        let interpolation = poly::point::interpolate(points);
-        assert_eq!(interpolation, poly)
-    }
-
-    #[test]
-    fn test_recover_overdetermined_poly() {
-        let poly = vec![g!(1 * G), g!(2 * G), g!(3 * G)];
-        let indicies = vec![
-            s!(1).public(),
-            s!(2).public(),
-            s!(3).public(),
-            s!(4).public(),
-            s!(5).public(),
-        ];
-        let points = indicies
-            .clone()
-            .into_iter()
-            .map(|index| {
-                (
-                    index,
-                    poly::point::eval(&poly, index.public())
-                        .normalize()
-                        .non_zero()
-                        .unwrap(),
-                )
-            })
-            .collect::<Vec<_>>();
-
-        let interpolation = poly::point::interpolate(points);
-
-        let (interpolated_coeffs, zero_coeffs) = interpolation.split_at(poly.len());
-        let n_extra_points = indicies.len() - poly.len();
-        assert_eq!(
-            (0..n_extra_points)
-                .map(|_| Point::<Normal, Public, Zero>::zero().public().normalize())
-                .collect::<Vec<_>>(),
-            zero_coeffs.to_vec()
-        );
-        assert_eq!(interpolated_coeffs, poly);
-    }
-
-    #[test]
-    fn test_reconstruct_shared_secret() {
-        let indicies = vec![s!(1).public(), s!(2).public(), s!(3).public()];
-        let scalar_poly = vec![s!(42), s!(53), s!(64)];
-
-        let secret_shares: Vec<_> = indicies
-            .clone()
-            .into_iter()
-            .map(|index| (index, poly::scalar::eval(&scalar_poly, index)))
-            .collect();
-
-        let reconstructed_secret = poly::scalar::interpolate_and_eval_poly_at_0(secret_shares);
-        assert_eq!(scalar_poly[0], reconstructed_secret);
-    }
 }
